requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
pydantic-settings==2.8.1
SQLAlchemy==2.0.38
tronpy==0.5.0
alembic==1.15.1
asyncpg==0.30.0
black==25.1.0
fastapi-pagination==0.12.34